backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cerebras_cloud_sdk==1.50.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
dotenv==0.9.9
exa-py==1.15.6
exceptiongroup==1.3.0
fastapi==0.118.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
openai==2.1.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
